requirements.txt
langchain==0.1.17
langchain_openai==0.1.6
openai==1.28.0
pandas==2.2.2
streamlit==1.34.0
python-dotenv==0.20.0
